+++ test/integration/cfg_test.lua
@@ -27,6 +27,8 @@ group.test_defaults = function(g)
         stats_driver = stats.get_default_driver(),
         stats_quantiles = false,
         stats_quantile_tolerated_error = 1e-3,
+        stats_quantile_age_buckets_count = 2,
+        stats_quantile_max_age_time = 60,
     })
 end
 
@@ -99,4 +101,48 @@ group.test_gh_284_preset_stats_quantile_tolerated_error_is_preserved = function(
     t.assert_equals(cfg.stats, true)
     t.assert_equals(cfg.stats_quantile_tolerated_error, 1e-4,
         'Preset stats_quantile_tolerated_error presents')
+end
+
+group.test_gh_284_preset_stats_quantile_age_buckets_count_is_preserved = function(g)
+    -- Arrange some cfg values so test case will not depend on defaults.
+    local cfg = g.cluster:server('router'):eval(
+        "return require('crud').cfg(...)",
+        {{ stats = false }})
+    t.assert_equals(cfg.stats, false)
+
+    -- Set stats_age_buckets_count.
+    local cfg = g.cluster:server('router'):eval(
+        "return require('crud').cfg(...)",
+        {{ stats_quantile_age_buckets_count = 3 }})
+    t.assert_equals(cfg.stats_quantile_age_buckets_count, 3)
+
+    -- Set another cfg parameter, assert preset stats_quantile_age_buckets_count presents.
+    local cfg = g.cluster:server('router'):eval(
+        "return require('crud').cfg(...)",
+        {{ stats = true }})
+    t.assert_equals(cfg.stats, true)
+    t.assert_equals(cfg.stats_quantile_age_buckets_count, 3,
+        'Preset stats_quantile_age_buckets_count presents')
+end
+
+group.test_gh_284_preset_stats_quantile_max_age_time_is_preserved = function(g)
+    -- Arrange some cfg values so test case will not depend on defaults.
+    local cfg = g.cluster:server('router'):eval(
+        "return require('crud').cfg(...)",
+        {{ stats = false }})
+    t.assert_equals(cfg.stats, false)
+
+    -- Set stats_age_buckets_count.
+    local cfg = g.cluster:server('router'):eval(
+        "return require('crud').cfg(...)",
+        {{ stats_quantile_max_age_time = 30 }})
+    t.assert_equals(cfg.stats_quantile_max_age_time, 30)
+
+    -- Set another cfg parameter, assert preset stats_quantile_max_age_time presents.
+    local cfg = g.cluster:server('router'):eval(
+        "return require('crud').cfg(...)",
+        {{ stats = true }})
+    t.assert_equals(cfg.stats, true)
+    t.assert_equals(cfg.stats_quantile_max_age_time, 30,
+        'Preset stats_quantile_max_age_time presents')
 end